Move FeatureAssert to testutils module
